requirements.txt
# Requirements for running orchestrate.py on the host machine
docker>=7.0.0
psycopg2-binary>=2.9.0
